AI-Othello-Arena/AI-Othello-Arena-web/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
numpy
